backend/requirements.txt
Pillow>=9.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
tensorflow>=2.10.0
numpy>=1.21.0
python-multipart>=0.0.5